Move reuseable parts of timeHistoryAnalyser to utils.anahistory
Huge commit but very few changes in the end.

See PR #112
